requirements.txt
amqp==5.0.1
asgiref==3.3.0
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.3
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-filter==2.4.0
django-webpack-loader==0.7.0
djangorestframework==3.12.1
idna==2.10
importlib-metadata==2.0.0
kombu==5.0.2
prompt-toolkit==3.0.8
psycopg2==2.8.6
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
